ski-resort-ranking/requirements.txt
pandas
requests
python-dotenv
kafka-python
snowflake-connector-python[pandas]
